refactor: replace maxAge of cookies to constant from number refactor: change to absolute import refactor: replace UserDto to separate file
