+++ src/modules/auth/auth.module.ts
@@ -8,10 +8,10 @@ import { MailModule } from 'src/modules/mail/mail.module';
 import { UserModule } from 'src/modules/user/user.module';
 import { SIGNIN_EXPIRATION_TIME } from 'src/utils/generalConstants';
 
-import { OtpService } from './services/otp.service';
-import { SuperAdminSeedService } from './services/super-admin-seed.service';
-import { JwtStrategy } from './strategies/jwt.strategy';
-import { LocalStrategy } from './strategies/local.strategy';
+import { OtpService } from 'src/modules/auth/services/otp.service';
+import { SuperAdminSeedService } from 'src/modules/auth/services/super-admin-seed.service';
+import { JwtStrategy } from 'src/modules/auth/strategies/jwt.strategy';
+import { LocalStrategy } from 'src/modules/auth/strategies/local.strategy';
 
 @Module({
   imports: [
