+++ src/modules/auth/strategies/local.strategy.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-local';
-import { UserDto } from 'src/modules/auth/dto/sign-in.response.dto';
+import { UserDto } from 'src/modules/auth/dto/user.response.dto';
 import { AuthService } from 'src/modules/auth/services/auth.service';
 
 @Injectable()
