+++ src/modules/google-auth/dto/google-auth.response.dto.ts
@@ -1,6 +1,5 @@
 import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
-
-import { EGoogLeAuthAction } from '../enums/google-auth-action.enum';
+import { EGoogLeAuthAction } from 'src/modules/google-auth/enums/google-auth-action.enum';
 
 export class GoogleAuthResponseDto {
   @ApiProperty()
