+++ src/modules/repository/services/user.repository.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@nestjs/common';
+import { UserEntity } from 'src/entities/user.entity';
 import { DataSource, Repository } from 'typeorm';
-
-import { UserEntity } from '../../../entities/user.entity';
 
 @Injectable()
 export class UserRepository extends Repository<UserEntity> {
