+++ src/modules/user/user.module.ts
@@ -1,7 +1,6 @@
 import { Module } from '@nestjs/common';
-
-import { UserController } from './user.controller';
-import { UserService } from './user.service';
+import { UserController } from 'src/modules/user/user.controller';
+import { UserService } from 'src/modules/user/user.service';
 
 @Module({
   imports: [],
