+++ src/utils/generalConstants.ts
@@ -6,6 +6,7 @@ export const TEST_MODULE_ROUTE = 'testEntity';
 export const SIGNIN_EXPIRATION_TIME = '60m';
 
 export const GOOGLE_REGISTR_RES = '/faq_team_fe/confirm-credentials?token=';
+export const COOKIES_AGE = 2592000000; // 30 days in milliseconds
 
 export const swgBuilderLabels = {
   title: 'Black-circle',
